build: split logging test helpers out and drop unused msp testtools
Two more items from the review of #1693.

Split platform/common/services/logging/logger.go. It mixed 13 production
symbols with test-only helpers, so the whole file had to stay in the coverage
denominator. The test helpers move to logging_test_utils.go, which the filter
excludes, leaving logger.go as production code only.

Four symbols moved, all test-only: NewTestLogger, the Recorder and Option
aliases for floggingtest, and Named. Named looks like production API but
takes a *floggingtest.RecordingCore and is called only from a _test.go, so it
belongs with the others. logger.go no longer imports testing or floggingtest.

Remove platform/fabric/core/msp/mgmt/testtools. The package was self-testing
dead code: its only caller was its own config_test.go, and mgmt_test.go --
the one place that would naturally use it -- defines a private
LoadMSPSetupForTesting of its own and never imports the package. Checked
against LFDT-Panurus/panurus first: code search returns no hits for
LoadMSPSetupForTesting or the package path, against 3 hits for the same
symbol in this repo as a control.

make checks passes on all four modules; the logging, diag and msp test
packages pass.
