--- platform/fabric/core/msp/mgmt/testtools/config.go
@@ -1,48 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package msptesttools
-
-import (
-	"path/filepath"
-
-	"github.com/hyperledger/fabric-lib-go/bccsp/factory"
-
-	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/msp"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/msp/mgmt"
-)
-
-// LoadMSPSetupForTesting sets up the local MSP
-// and a chain MSP for the default chain
-func LoadMSPSetupForTesting() error {
-	dir, err := filepath.Abs("../../testdata/sampleconfig")
-	if err != nil {
-		return err
-	}
-
-	return LoadMSPSetupForTestingFromDir(dir)
-}
-
-// LoadMSPSetupForTestingFromDir sets up the local MSP
-// and a chain MSP for the default chain using the given directory
-func LoadMSPSetupForTestingFromDir(dir string) error {
-	conf, err := msp.GetLocalMspConfig(dir, nil, "SampleOrg")
-	if err != nil {
-		return err
-	}
-
-	err = mgmt.GetLocalMSP(factory.GetDefault()).Setup(conf)
-	if err != nil {
-		return err
-	}
-
-	err = mgmt.GetManagerForChain("testchannelid").Setup([]msp.MSP{mgmt.GetLocalMSP(factory.GetDefault())})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
